refactor: vfx components and state management
- Mark tasks 1.0-3.0 as completed in REFACTORING_TASKS.md
- Document remaining tasks 4.0-6.0 with specific instructions and effort estimates
- Create docs/refactoring/README.md as navigation and quick reference
- Include branch naming conventions and verification steps
- Add status table showing completed vs. remaining work

Completed tasks this session:
- Task 1.0: Consolidate AlbumFilters interface
- Task 2.0: Eliminate wrapper components
- Task 3.0: Consolidate keyboard shortcuts

Ready for next developer to pick up Task 4.0

Refactor useVisualEffectsState to use useLocalStorage hook

Abstract visualizer patterns into useCanvasVisualizer hook

Extract VisualEffectsMenu styled components
