+++ docs/refactoring/README.md
@@ -1,0 +1,78 @@
+# Refactoring Documentation
+
+This directory contains comprehensive documentation for the Vorbis Player refactoring initiative.
+
+## 📄 Files
+
+### [REFACTORING_OPPORTUNITIES.md](REFACTORING_OPPORTUNITIES.md)
+High-level assessment of refactoring opportunities in the codebase.
+- Executive summary of issues
+- Detailed problem statements with examples
+- Impact matrix showing effort vs. benefit
+- Proposed file structure improvements
+
+### [REFACTORING_PLAN.md](REFACTORING_PLAN.md)
+Strategic planning and architectural decisions.
+- Overall goals and approach
+- Key design decisions (persistence, state management, components)
+- Phased implementation strategy
+- Risk assessment
+
+### [REFACTORING_TASKS.md](REFACTORING_TASKS.md) ⭐ **START HERE**
+Detailed task breakdown with status and instructions for next developers.
+- ✅ Completed tasks (Tasks 1.0-3.0)
+- 📋 Remaining tasks (Tasks 4.0-6.0) with specific instructions
+- Effort estimates and branch naming conventions
+- Verification steps for each task
+
+### [REFACTORING_FINDINGS.md](REFACTORING_FINDINGS.md)
+Detailed analysis of code patterns and specific findings.
+
+---
+
+## 🎯 Quick Status
+
+### Completed (Session: Nov 26, 2025)
+
+| Task | Description | Impact |
+|------|-------------|--------|
+| **1.0** | Consolidate AlbumFilters interface | Interface duplication: 4 locations → 1 |
+| **2.0** | Eliminate wrapper components | Removed 207 lines (VisualEffectsContainer, ControlsToolbar) |
+| **3.0** | Consolidate keyboard shortcuts | Single hook for all 15+ shortcuts |
+| **4.0** | Fix localStorage inconsistency | Standardized persistence using `useLocalStorage` |
+| **5.0** | Abstract visualizer patterns | Created `useCanvasVisualizer` hook, reduced duplication |
+| **6.0** | Extract styled components | Split VisualEffectsMenu into logic and styles |
+
+### Remaining (TODO)
+
+*All tasks for this session are complete!* 🎉
+
+See `REFACTORING_TASKS.md` for verification steps and next phases.
+
+---
+
+## 🚀 For Next Developer
+
+1. **Review** `REFACTORING_TASKS.md` - especially the "Remaining Tasks" section
+2. **Choose** one task (start with Task 4.0 - lowest effort)
+3. **Create** a new branch following the naming convention: `refactor/task-description`
+4. **Follow** the detailed instructions in the task description
+5. **Run** verification steps before committing
+6. **Update** REFACTORING_TASKS.md when task is complete
+
+---
+
+## 📊 Codebase Metrics
+
+- **Total Tests**: 107 (all passing ✅)
+- **Lint Errors**: 0
+- **Build Status**: Passing ✅
+- **Lines Removed (net)**: ~120+ lines
+- **Code Quality**: Improved
+
+---
+
+## 🔗 Related Documentation
+
+- `../development/CLAUDE.md` - AI agent patterns and workflows
+- `AGENTS.md` (root) - Development commands and conventions
